+++ src/resourcetypes/index.js
@@ -7,3 +7,4 @@ angular.module('md.resourcetypes', ['md.forms', 'ui.router']);
 
 require('./states');
 require('./controllers');
+require('./services');