+++ demos/full/controllers.js
@@ -4,12 +4,7 @@
     this.features = resource.items;
   }
 
-  function CollapseCtrl($scope) {
-    $scope.isCollapsed = true;
-  }
-
   angular.module('full')
-    .controller('FeaturesCtrl', FeaturesCtrl)
-    .controller('CollapseCtrl', CollapseCtrl);
+    .controller('FeaturesCtrl', FeaturesCtrl);
 
 })();