+++ src/configurator/index.js
@@ -1,4 +1,8 @@
 'use strict';
 
+var angular = require('angular');
+angular.module('md.config', []);
+
 require('./services');
+require('./directives');