+++ src/forms/directives.js
@@ -2,7 +2,6 @@ function FormCtrl(MdSchemas, MdForms) {
   this.model = this.mdModel;
   this.schema = MdSchemas.get(this.mdSchema);
   this.form = MdForms.get(this.mdForm);
-  console.debug('model48', this.model);
 }
 
 
@@ -21,30 +20,5 @@ function Form() {
   };
 }
 
-
-function ResourceTypeCtrl(Restangular) {
-  var self = this;
-  var url = '/api/' + this.mdResource;
-  Restangular.one(url).get().then(function (resource) {
-    self.resource = resource;
-  });
-}
-
-
-function ResourceType() {
-  return {
-    restrict: "E",
-    templateUrl: "/forms/templates/resource.html",
-    scope: {
-      mdResource: '@mdResource'
-    },
-    controller: ResourceTypeCtrl,
-    controllerAs: 'ctrl',
-    bindToController: true
-  };
-}
-
-
 angular.module("md.forms")
-  .directive("mdForm", Form)
-  .directive("mdResourceType", ResourceType);
+  .directive("mdForm", Form);