+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## Unreleased
 
-- CRUD
+
 
 - Provide ``md-init`` directive as an attribute that can point to the 
 URL of a JSON file that bootstraps everything.